+++ assets/js/base/hooks/index.js
@@ -4,4 +4,7 @@ export * from './use-position-relative-to-viewport';
 export * from './use-previous';
 export * from './use-shallow-equal';
 export * from './use-throw-error';
-export * from './style-attributes';
+export * from './use-spacing-props';
+export * from './use-typography-props';
+export * from './use-color-props';
+export * from './use-border-props';
