+++ assets/js/base/utils/index.js
@@ -7,3 +7,4 @@ export * from './get-valid-block-attributes';
 export * from './product-data';
 export * from './derive-selected-shipping-rates';
 export * from './get-icons-from-payment-methods';
+export * from './parse-style';
